+++ zarrs/src/array/array_async_sharded_readable_ext.rs
@@ -86,7 +86,7 @@ impl AsyncArrayShardedReadableExtCache {
         Self {
             array_is_sharded: array.is_sharded(),
             array_is_exclusively_sharded: array.is_exclusively_sharded(),
-            subchunk_grid,
+            subchunk_grid: subchunk_grid.clone(),
             cache: Arc::new(async_lock::Mutex::new(HashMap::default())),
         }
     }
@@ -796,12 +796,12 @@ mod tests {
                 ])
             );
             assert_eq!(
-                array.effective_subchunk_shape(),
-                Some(vec![
+                array.partial_decode_granularity(&[0, 0, 0])?,
+                vec![
                     NonZeroU64::new(2).unwrap(),
                     NonZeroU64::new(1).unwrap(),
                     NonZeroU64::new(3).unwrap()
-                ])
+                ]
             ); // NOTE: transposed
             assert_eq!(subchunk_grid.grid_shape(), &[8, 16, 3]);
         } else {