+++ zarrs/tests/array_sync.rs
@@ -7,7 +7,7 @@ use std::sync::Arc;
 use zarrs::array::{Array, ArrayBuilder, ArrayBytes, ArraySubset, FillValue, data_type};
 use zarrs::storage::store::MemoryStore;
 use zarrs_codec::{
-    ArrayBytesDecodeIntoTarget, ArrayBytesFixedDisjointView, ArrayCodecTraits, CodecOptions,
+    ArrayBytesDecodeIntoTarget, ArrayBytesFixedDisjointView, ArrayToBytesCodecTraits, CodecOptions,
 };
 
 #[allow(clippy::single_range_in_vec_init)]
@@ -120,7 +120,10 @@ fn array_sync_read_uncompressed() -> Result<(), Box<dyn std::error::Error>> {
 
     let chunk_shape = array.chunk_shape(&vec![0; array.dimensionality()])?;
     assert_eq!(
-        array.codecs().partial_decode_granularity(&chunk_shape),
+        array
+            .codecs()
+            .partial_decode_granularity(&chunk_shape)
+            .unwrap(),
         [NonZeroU64::new(2).unwrap(); 2]
     );
 
@@ -155,7 +158,10 @@ fn array_sync_read_shard_compress() -> Result<(), Box<dyn std::error::Error>> {
 
     let chunk_shape = array.chunk_shape(&vec![0; array.dimensionality()])?;
     assert_eq!(
-        array.codecs().partial_decode_granularity(&chunk_shape),
+        array
+            .codecs()
+            .partial_decode_granularity(&chunk_shape)
+            .unwrap(),
         [NonZeroU64::new(1).unwrap(); 2]
     );